moved all EntityType related extensions to own file
