+++ game/engine/src/main/kotlin/de/gleex/pltcmd/game/engine/attributes/DestinationAttribute.kt
@@ -1,8 +1,5 @@
 package de.gleex.pltcmd.game.engine.attributes
 
-import de.gleex.pltcmd.game.engine.entities.Movable
-import de.gleex.pltcmd.game.engine.extensions.GameEntity
-import de.gleex.pltcmd.game.engine.extensions.getAttribute
 import de.gleex.pltcmd.model.world.coordinate.Coordinate
 import org.hexworks.amethyst.api.Attribute
 import org.hexworks.cobalt.databinding.api.extension.createPropertyFrom
@@ -18,14 +15,3 @@ class DestinationAttribute(initialDestination: Coordinate? = null) : Attribute {
             coordinateProperty.updateValue(value)
         }
 }
-
-var GameEntity<Movable>.destination: Maybe<Coordinate>
-    get() = getAttribute(DestinationAttribute::class).coordinate
-    set(value) {
-        findAttribute(DestinationAttribute::class).map {
-            it.coordinate = value
-        }
-    }
-
-val GameEntity<Movable>.hasNoDestination: Boolean
-    get() = destination.isEmpty()
