+++ game/engine/src/main/kotlin/de/gleex/pltcmd/game/engine/entities/EntityTypes.kt
@@ -1,13 +1,7 @@
 package de.gleex.pltcmd.game.engine.entities
 
+import de.gleex.pltcmd.game.engine.entities.types.Movable
 import org.hexworks.amethyst.api.base.BaseEntityType
-import org.hexworks.amethyst.api.entity.EntityType
-
-/** Type marker for entities that have the PositionAttribute */
-interface Positionable: EntityType
-
-/** Type marker for entities that are [Positionable] and have the DestinationAttribute */
-interface Movable: Positionable
 
 /** Represents an element in an army. */
 object ElementType : BaseEntityType("element", "A movable element."), Movable
